Moved Actor Ls from node api to plumbing
