+++ api/api.go
@@ -3,7 +3,6 @@ package api
 
 // API is the user interface to a Filecoin node.
 type API interface {
-	Actor() Actor
 	Address() Address
 	Client() Client
 	Daemon() Daemon
